feat html structure
This is fundamental to initiate the project in order
to have a basic strucutre from which to start and
then write all the code. This has been done by
using the html:5 shortcut since I already know all
the html boilerplate parts and also added the h1 title
that says "Odin Recipes" to indicate the main
purpose of the webpage.