feat new positioning and styling
This commit adds margins to texts to separate
and help readability. It also limitates p
elements max-width to reduce the amoun of inline
space they take, and this also enhances
the text hierarchy. Finally, created footer section
at the end of lasagna recipe since it was necesary
to separate the instructions for preparation
from the last indication.